back/requirements.txt
langchain==0.1.13
langchain-openai==0.1.1
langchainhub==0.1.15
unstructured==0.12.6
Markdown==3.6
markdownify==0.11.6
langserve==0.0.51
fastapi==0.110.0
python-dotenv==1.0.1
jsonschema==4.21.1
requests==2.31.0
beautifulsoup4==4.12.3
uvicorn==0.29.0
langsmith==0.1.31
azure-storage-blob==12.19.1
azure-search-documents==11.4.0
azure-identity==1.16.0
unstructured==0.12.6
